+++ apps/github-search/package.json
@@ -17,6 +17,7 @@
     "@chakra-ui/react": "^1.3.4",
     "@emotion/react": "^11.1.5",
     "@emotion/styled": "^11.1.5",
+    "@howardism/theme": "^0.1.0",
     "framer-motion": "^3.10.3",
     "next": "10.0.8",
     "react": "17.0.1",
