+++ packages/theme/package.json
@@ -4,6 +4,7 @@
   "main": "index.ts",
   "license": "MIT",
   "dependencies": {
-    "@chakra-ui/react": "^1.3.4"
+    "@chakra-ui/react": "^1.3.4",
+    "react": "^17.0.1"
   }
 }
